+++ previous-versions/files/src/codatfiles/models/operations/list_files.py
@@ -3,8 +3,10 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
 from ..shared import file as shared_file
+from ..shared import schema as shared_schema
+from codatfiles import utils
+from dataclasses_json import Undefined, dataclass_json
 from typing import Optional
 
 
@@ -16,15 +18,36 @@ class ListFilesRequest:
 
 
 
+@dataclass_json(undefined=Undefined.EXCLUDE)
+
+@dataclasses.dataclass
+class ListFilesErrorMessage:
+    r"""One or more of the resources you referenced could not be found.
+    This might be because your company or data connection id is wrong, or was already deleted.
+    """
+    can_be_retried: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('canBeRetried'), 'exclude': lambda f: f is None }})
+    correlation_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('correlationId'), 'exclude': lambda f: f is None }})
+    detailed_error_code: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('detailedErrorCode'), 'exclude': lambda f: f is None }})
+    error: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('error'), 'exclude': lambda f: f is None }})
+    service: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('service'), 'exclude': lambda f: f is None }})
+    status_code: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('statusCode'), 'exclude': lambda f: f is None }})
+    
+
+
+
 
 @dataclasses.dataclass
 class ListFilesResponse:
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
+    error_message: Optional[ListFilesErrorMessage] = dataclasses.field(default=None)
+    r"""One or more of the resources you referenced could not be found.
+    This might be because your company or data connection id is wrong, or was already deleted.
+    """
     files: Optional[list[shared_file.File]] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    schema: Optional[shared_schema.Schema] = dataclasses.field(default=None)
+    r"""Your API request was not properly authorized."""